Refactored file structure for drawables
